+++ src/lib/Layout/PanelSideBarLayout/PanelSideBar/PanelSideBarItem.tsx
@@ -11,16 +11,43 @@ export interface PanelSideBarItemProps<TPanelItemId extends string, TPanelItem> 
   depth?: number;
   active?: boolean;
   isParentHidden?: boolean;
+  isIconShownOnSidebarCollapse: boolean;
 }
+
+const PanelSidebarItemNavLink = <TPanelItemId extends string, TPanelItem>({
+  item,
+  collapsedWithIcon,
+  className,
+}: {
+  item: PanelItem<TPanelItemId, TPanelItem>;
+  collapsedWithIcon?: boolean;
+  className?: string;
+}) => {
+  const { icon, title, onSidebarCollapseOptions } = item;
+  const panelIconClassName = collapsedWithIcon ? "ms-1 me-3 p-1" : "me-2";
+  const displayIcon = icon || (collapsedWithIcon && onSidebarCollapseOptions?.fallbackIcon);
+
+  return (
+    <span className={className}>
+      {displayIcon && <FontAwesomeIcon icon={displayIcon} className={panelIconClassName} />}
+      {!collapsedWithIcon || !displayIcon ? title : ""}
+    </span>
+  );
+};
 
 // eslint-disable-next-line complexity
 const PanelSideBarItem = <TPanelItemId extends string, TPanelItem>(props: PanelSideBarItemProps<TPanelItemId, TPanelItem>) => {
-  const { depth = 0, children: item, isParentHidden = false } = props;
-  const { LinkRenderer, toggledMenuItemIds, toggleMenuItem, hiddenMenuItemIds } = usePanelSideBarContext<TPanelItemId, TPanelItem>();
-  const hasitem = !!item.children?.length;
-  const isActive = (hasitem && item.children && hasActiveChildren(item.children)) || item.active;
+  const { depth = 0, children: item, isParentHidden = false, isIconShownOnSidebarCollapse } = props;
+  const { LinkRenderer, toggledMenuItemIds, toggleMenuItem, hiddenMenuItemIds, isSidebarOpen } = usePanelSideBarContext<
+    TPanelItemId,
+    TPanelItem
+  >();
+
+  const hasItems = !!item.children?.length;
+  const isActive = (hasItems && item.children && hasActiveChildren(item.children)) || item.active;
   const isOpen = toggledMenuItemIds?.includes(item.id);
   const scrollToActiveItemRef = useRef<HTMLDivElement>(null);
+  const collapsedWithIcon = isIconShownOnSidebarCollapse && !isSidebarOpen;
 
   useEffect(() => {
     if (scrollToActiveItemRef.current && isActive) {
@@ -33,28 +60,33 @@ const PanelSideBarItem = <TPanelItemId extends string, TPanelItem>(props: PanelS
       <NavItem
         hidden={isParentHidden || hiddenMenuItemIds.includes(item.id)}
         onClick={() => {
-          if (hasitem && !item.collapseIconOnly) {
+          if (hasItems && !item.collapseIconOnly) {
             toggleMenuItem(item.id);
           }
         }}
         className={classNames({ "menu-open": isOpen, active: isActive })}
-        style={{ paddingLeft: depth ? `${depth + 1}rem` : undefined }}
+        style={{ paddingLeft: !collapsedWithIcon && depth ? `${depth + 1}rem` : undefined }}
       >
         <div ref={scrollToActiveItemRef}>
-          {hasitem ? (
+          {hasItems ? (
             <div className={classNames("d-flex flex-row", { "justify-content-between": item.collapseIconOnly })}>
               {item.collapseIconOnly && (
                 <LinkRenderer item={item}>
-                  <span className="nav-link">
-                    {item.icon && <FontAwesomeIcon icon={item.icon} className="me-2" />}
-                    {item.title}
-                  </span>
+                  <PanelSidebarItemNavLink<TPanelItemId, TPanelItem>
+                    className="nav-link"
+                    item={item}
+                    collapsedWithIcon={collapsedWithIcon}
+                  />
                 </LinkRenderer>
               )}
 
               <a
                 role="button"
-                className={classNames("nav-link", { "w-100": !item.collapseIconOnly }, { "dropdown-toggle": hasitem })}
+                className={classNames(
+                  "nav-link",
+                  { "w-100": !item.collapseIconOnly },
+                  { "dropdown-toggle": hasItems && !collapsedWithIcon },
+                )}
                 onClick={() => {
                   if (item.collapseIconOnly) {
                     toggleMenuItem(item.id);
@@ -62,26 +94,20 @@ const PanelSideBarItem = <TPanelItemId extends string, TPanelItem>(props: PanelS
                 }}
               >
                 {!item.collapseIconOnly && (
-                  <span>
-                    {item.icon && <FontAwesomeIcon className="me-2" icon={item.icon} />}
-                    {item.title}
-                  </span>
+                  <PanelSidebarItemNavLink<TPanelItemId, TPanelItem> item={item} collapsedWithIcon={collapsedWithIcon} />
                 )}
               </a>
             </div>
           ) : (
             <>
               <LinkRenderer item={item}>
-                <span className="nav-link">
-                  {item.icon && <FontAwesomeIcon icon={item.icon} className="me-2" />}
-                  {item.title}
-                </span>
+                <PanelSidebarItemNavLink<TPanelItemId, TPanelItem> className="nav-link" item={item} collapsedWithIcon={collapsedWithIcon} />
               </LinkRenderer>
             </>
           )}
         </div>
       </NavItem>
-      {hasitem && (
+      {hasItems && (
         <Collapse isOpen={isOpen} navbar className={classNames("item-menu", { "mb-1": isOpen })}>
           {item.children?.map((childItem) => (
             <PanelSideBarItem
@@ -90,6 +116,7 @@ const PanelSideBarItem = <TPanelItemId extends string, TPanelItem>(props: PanelS
               depth={depth + 1}
               active={item.active}
               isParentHidden={hiddenMenuItemIds.includes(item.id)}
+              isIconShownOnSidebarCollapse={isIconShownOnSidebarCollapse}
             />
           ))}
         </Collapse>