+++ src/lib/Layout/PanelSideBarLayout/PanelSideBar/PanelSideBarToggle.tsx
@@ -6,10 +6,11 @@ import classNames from "classnames";
 
 interface PanelSideBarToggleProps extends ButtonProps {
   toggled: boolean;
+  isIconShownOnSidebarCollapse: boolean;
 }
 
 export const PanelSideBarToggle = (props: PanelSideBarToggleProps) => {
-  const { toggled, ...buttonProps } = props;
+  const { toggled, isIconShownOnSidebarCollapse, ...buttonProps } = props;
   const { theme } = usePanelSideBarContext();
 
   return (
@@ -19,6 +20,7 @@ export const PanelSideBarToggle = (props: PanelSideBarToggleProps) => {
         { "side-nav-toggle-dark": theme == "dark" },
         { "side-nav-toggle-light": theme == "light" },
         { "side-nav-toggle-blue": theme == "blue" },
+        { "show-icons": isIconShownOnSidebarCollapse },
       )}
       id="side-nav-toggle"
       color="primary"