+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added the pkg.pr.new workflow.
+- `onSidebarCollapseOptions` to customize the panel item behavior on sidebar collapsing: keep showing icons enabling `showIcon`, use a fallback icon setting `fallbackIcon`. It works:
+  - whether tiles are rendered via `renderFirstItemsLevelAsTiles`
+  - the panel item has any `children`
 
 ## [4.2.0] - 2024-10-08